requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
google-generativeai>=0.3.0
reportlab>=4.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
